python-project2-1/requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
pymongo==4.3.3
Werkzeug==2.2.2
